+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Version 0.10.7
 * Added door state to blueprints (only interpreted on "direct" build)
 * Added ward state (active/inactive) to blueprints (only interpreted on "direct" build)
+* Added config value for the smoothing option of the terrain tools
 
 # Version 0.10.6
 * Fixed undo not removing pieces from blueprint selections